+++ library/command-information.js
@@ -1,12 +1,7 @@
-const Command = require('./command');
+const AdministrationCommand = require('../commands/administration');
+const CoreCommand = require('../commands/core');
 const FunCommand = require('../commands/fun');
-const HelpCommand = require('../commands/help');
 const UsefulCommand = require('../commands/useful');
-const PurgeCommand = require('../commands/purge');
-const RulesCommand = require('../commands/rules');
-
-const command = new Command();
-const usefulCommand = new UsefulCommand();
 
 /**
  * TODO: 
@@ -17,22 +12,22 @@ let commands = [
   {
     command: 'add [Integers]',
     description: 'Adds together _integers_ passed as arguments.',
-    execution: { add: (args, msg) => usefulCommand.add(args, msg) },
+    execution: { add: (args, msg) => UsefulCommand.add(args, msg) },
   },
   {
     command: 'gitProfile [Username]',
     description: 'Retrieves any public github profile.',
-    execution: { gitProfile: (args, msg) => usefulCommand.gitProfile(args, msg) },
+    execution: { gitProfile: (args, msg) => UsefulCommand.gitProfile(args, msg) },
   },
   {
     command: 'help',
     description: 'Displays this message',
-    execution: { help: (args, msg, commands) => HelpCommand.help(args, msg, commands) },
+    execution: { help: (args, msg, commands) => CoreCommand.help(args, msg, commands) },
   },
   {
     command: 'lmgtfy [Search Term]',
     description: 'When someone is being...lazy...?',
-    execution: { lmgtfy: (args, msg) => usefulCommand.lmgtfy(args, msg) },
+    execution: { lmgtfy: (args, msg) => UsefulCommand.lmgtfy(args, msg) },
   },
   {
     command: 'magic8ball [Question]',
@@ -42,12 +37,12 @@ let commands = [
   {
     command: 'purge',
     description: 'Purges the channel it is called within. Restricted to Board Members and Administrators.',
-    execution: { purge: (args, msg) => PurgeCommand.purge(args, msg) },
+    execution: { purge: (args, msg) => AdministrationCommand.purge(args, msg) },
   },
   {
     command: 'rules',
     description: 'List the rules for the CTC Discord server.',
-    execution: { rules: (args, msg) => RulesCommand.rules(args, msg) },
+    execution: { rules: (args, msg) => CoreCommand.rules(args, msg) },
   },
   {
     command: 'say [Phrase]',
@@ -57,17 +52,17 @@ let commands = [
   {
     command: 'search [Search Term]',
     description: 'Searches the web for the passed query and return the top result.',
-    execution: { search: (args, msg) => usefulCommand.search(args, msg) },
+    execution: { search: (args, msg) => UsefulCommand.search(args, msg) },
   },
   {
     command: 'source',
     description: 'Retrieves the hackBot\'s github repository.',
-    execution: { source: (args, msg) => command.source(args, msg) },
+    execution: { source: (args, msg) => CoreCommand.source(args, msg) },
   },
   {
     command: 'weather [Location]',
     description: 'Provide City and State, or City and Country to get current temperature.',
-    execution: { weather: (args, msg) => usefulCommand.weather(args, msg) },
+    execution: { weather: (args, msg) => UsefulCommand.weather(args, msg) },
   },
   {
     command: 'xmas',
