+++ library/command.js
@@ -1,4 +1,11 @@
-
+/**
+ * This is no longer doing anything as far as I know.
+ * This should provide the structure that each commands should have.
+ * The structure should be based off of the command-information.js
+ * After this is implemented we will be able to remove command-information.js
+ * 
+ * @class Command
+ */
 class Command {
   constructor() {
